LayoutTests/imported/w3c/web-platform-tests/intersection-observer/multiple-targets-expected.txt


PASS One observer with multiple targets. 
PASS First rAF. 
FAIL document.scrollingElement.scrollTop = 150 assert_equals: Four notifications. expected 4 but got 3
FAIL document.scrollingElement.scrollTop = 10000 assert_equals: Six notifications. expected 6 but got 3
FAIL document.scrollingElement.scrollTop = 0 assert_equals: Nine notifications. expected 9 but got 3
